# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Match": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"scoredHomeTeam": {
						"type": "number",
						"format": "double"
					},
					"scoredAwayTeam": {
						"type": "number",
						"format": "double"
					},
					"homeTeam": {
						"$ref": "#/components/schemas/Team"
					},
					"awayTeam": {
						"$ref": "#/components/schemas/Team"
					},
					"tournament": {
						"$ref": "#/components/schemas/Tournament"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"date",
					"scoredHomeTeam",
					"scoredAwayTeam",
					"homeTeam",
					"awayTeam",
					"tournament",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Team": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"homeMatches": {
						"items": {
							"$ref": "#/components/schemas/Match"
						},
						"type": "array"
					},
					"awayMatches": {
						"items": {
							"$ref": "#/components/schemas/Match"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"homeMatches",
					"awayMatches",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Tournament": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"matches": {
						"items": {
							"$ref": "#/components/schemas/Match"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"matches",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGetMatchesResp": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Match"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-typescript",
		"version": "1.1.1",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/matches": {
			"get": {
				"operationId": "GetMatches",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGetMatchesResp"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "date",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/matches/availabilities": {
			"get": {
				"operationId": "GetAvailabilities",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"items": {
												"type": "string"
											},
											"type": "array"
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "date",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}